Fix media projection permission dialog too wide when in split screen
When in split screen and large screen, the dialog was using almost the
entire width of the screen.

SysUIDialog defines the dialog width based on a resource dimension.
For small screens, it is MATCH_PARENT, and for larger screens a specific
value.
For the media projection permission dialog, the context passed to the
dialog was of the activity that was on one of the sides of the split
screen, which has the same width as a small screen.

The fix is to pass the application context instead to the dialog.

Test: MediaProjectionPermissionTest
Test: Manually - Put the device in split screen while on a large screen
                 and start screen sharing.
Fixes: 263801365
Change-Id: Ifd600b7ec54deeabe28f4762ce87975d97dbc89d