Remove platform build team from frameworks/base build definition owners.
The platform build team isn't equipped with the deep frameworks context to review changes. The Android.bp and Android.mk changes should be reviewed by folks who work on frameworks day-to-day.

In any case, there are members on the platform build team with Global Approver powers, in case we need overriding in the future.

Change-Id: I05324a25d0c6037b6240f209bd4b23a740d440fd
Test: N/A