Beef up logging in and near tryRemoveNotification
We're ending up removing a notification twice -- probably a race
between clear all and dismissing or cancelling one of the notifications
hit by the clear all.

Long term, we want to figure out how this is happening, which means
catching the erroneous state changes earlier or at a higher level, so
this adds a bunch of logging around edge cases in the dismissal process
in NotifCollection.

Bug: 282945192
Test: presubmit
Change-Id: I452ae013b539208a26304caba17a3b8630f5e685