Checksum.java: fix links to fsverity documentation
These links were linking to the unrendered fsverity documentation file
on the old fscrypt development branch (not fsverity, and not the
mainline kernel).  This development branch was moved several months ago,
and therefore these links are dead now.

Fix these links to link to the rendered documentation for the mainline
kernel.

Test: comment only change
Change-Id: I3460caaf0efa235425cc1e1ae62e0bc2a4503177