Stop ellipsize in system server
System server don't have UI but doing ellipsize with real font with
hard coded font size and hard coded width. This is the only font
usage in the system server, so let's remove this for preparing
skiping font initialization in the system server.

In the system server, this method is used for loading label. During
loading labels, the hard coded font size 42 px and 1000px width has
been used. This is equivalent to about 41 characters in Roboto.

No behavior changes to the this public API users except for system
server.

Bug: 291239366
Test: Manually done by booting device (Pixel 5)
Change-Id: I71c940de4b76bf32c14ead66a7912dfb8f25a384