fix(non linear font scaling): battery icon in qqs status bar might be cut off
As the screenshot in b/291168770, when qqs status bar contains too much system icons and the container can not display all icons, the battery icon view might be cut off or even disappear. Since originally the battery icon view has higher priority and should stick to the right side of the container, we change the system icons container layout to ensure the battery icon view is placed properly, then using the remain space for the other system icons displaying.

Bug: 291168770
Test: manually - attached screenshot in bug
Change-Id: Id36f6cb0aa336dbbb4ca45c24879be3e4bc1e876