Send user activity when dream quits unexpectedly
If a dream quits unexpectedly after the device has been idle for longer
than the user's screen timeout, the device will go straight to a black
screen. This happens routinely when GMSCore or the dream app itself are
updated and is jarring for the user experience.

This CL sends a user activity signal to PowerManager when we see a
dream quit unexpectedly so that instead the device goes to keyguard and
will timeout back to dreaming after a short wait. This is the same
behavior as if a dream quit unexpectedly before the user activity
timeout expired and allows for a graceful recovery.

Bug: 286777009
Test: atest DreamControllerTest
Test: manually verified that crashing the dreams app goes to keyguard
	instead of a black screen
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:e0f3cb21ef0f34d118d28d45587ab03efd9cac00)
Merged-In: I47c9561756ac27370b597a69a948892641ff47bb
Change-Id: I47c9561756ac27370b597a69a948892641ff47bb