Update mInsetsHint regardless of the nullity of mControl
Befroe this CL, we only update InsetsSourceProvider#mInsetsHint while
there is a control. However, the very first control target can get the
control before the mInsetsHint gets updated in onPostLayout. This CL
updates mInsetsHint once the source frame is updated, so that we pass
the correct insets hint to the constructor of InsetsSourceControl.

Fix: 289892083
Test: atest InsetsSourceProviderTest
Test: Switch to 3-button navigation bar and reboot the device.
      Make sure the back button is not shown on the lockscreen.
Change-Id: Ic876284e4b7954d961786ab0ae65794e7318cb98