Update to cold launch if the process is restarted during launching
The current common case is that the process of the resuming activity
is just unfrozen but its binder buffer is full (someone still sends
async binder call to it when it is frozen) and the process could
be killed because binder transaction failed. Then the process will
attach again so the launch type can be corrected to COLD to avoid
disturb the distribution of metrics.

Bug: 291599121
Test: atest ActivityMetricsLaunchObserverTests#testLaunchState
Change-Id: I26252410508d21aacacd35bd6cc305dc7a0fc9f9
(cherry picked from commit ef8d2943879bc74a56f4058afb38909b8c04b61e)